backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
aiosqlite
python-dotenv
httpx
sse-starlette 